+++ packages/postgres/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `@verse-bot/postgres` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Published package includes the Apache-2.0 license file.
+
 ## [0.1.0] - 2026-09-23
 
 ### Added